Creates a base client having the responsibility for holding the HttpClient
Removes complexity of duplicating HttpClient code for each new action
