+++ Digipost.Signature.Api.Client.Core/Digipost.Signature.Api.Client.Core.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Asice\DataTransferObjects\DocumentDataTransferObject.cs" />
     <Compile Include="Asice\DataTransferObjects\SenderDataTransferObject.cs" />
     <Compile Include="Asice\DocumentBundle.cs" />
+    <Compile Include="BaseClient.cs" />
     <Compile Include="Internal\DigipostAction.cs" />
     <Compile Include="Internal\IRequestContent.cs" />
     <Compile Include="Internal\LoggingHandler.cs" />
